backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Multi-user task management backend with JWT authentication"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlmodel>=0.0.16",
    "pydantic>=2.5.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.1.18",
    "psycopg2-binary>=2.9.7",
    "aiosqlite>=0.19.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
